YCSB workloads work.

+++ bpt.c
@@ -91,11 +91,11 @@ int height( node * root );
 int path_to_root( node * root, node * child );
 void print_leaves( node * root );
 void print_tree( node * root );
-void find_and_print(node * root, int key, bool verbose); 
-void find_and_print_range(node * root, int range1, int range2, bool verbose); 
-int find_range( node * root, int key_start, int key_end, bool verbose,
+void find_and_print(node * root, size_t key, bool verbose); 
+void find_and_print_range(node * root, size_t range1, size_t range2, bool verbose); 
+int find_range( node * root, size_t key_start, size_t key_end, bool verbose,
 		int returned_keys[], void * returned_pointers[]); 
-node * find_leaf( node * root, int key, bool verbose );
+node * find_leaf( node * root, size_t key, bool verbose );
 int cut( int length );
 
 // Insertion.
@@ -104,15 +104,15 @@ record * make_record(char *value);
 node * make_node( void );
 node * make_leaf( void );
 int get_left_index(node * parent, node * left);
-node * insert_into_leaf( node * leaf, int key, record * pointer );
-node * insert_into_leaf_after_splitting(node * root, node * leaf, int key, record * pointer);
+node * insert_into_leaf( node * leaf, size_t key, record * pointer );
+node * insert_into_leaf_after_splitting(node * root, node * leaf, size_t key, record * pointer);
 node * insert_into_node(node * root, node * parent, 
-		int left_index, int key, node * right);
+		int left_index, size_t key, node * right);
 node * insert_into_node_after_splitting(node * root, node * parent, int left_index, 
-		int key, node * right);
-node * insert_into_parent(node * root, node * left, int key, node * right);
-node * insert_into_new_root(node * left, int key, node * right);
-node * start_new_tree(int key, record * pointer);
+		size_t key, node * right);
+node * insert_into_parent(node * root, node * left, size_t key, node * right);
+node * insert_into_new_root(node * left, size_t key, node * right);
+node * start_new_tree(size_t key, record * pointer);
 
 // Deletion.
 
@@ -121,7 +121,7 @@ node * adjust_root(node * root,int worker_id);
 node * coalesce_nodes(node * root, node * n, node * neighbor, int neighbor_index, int k_prime, int worker_id);
 node * redistribute_nodes(node * root, node * n, node * neighbor, int neighbor_index, 
 		int k_prime_index, int k_prime);
-node * delete_entry( node * root, node * n, int key, void * pointer , int worker_id);
+node * delete_entry( node * root, node * n, size_t key, void * pointer , int worker_id);
 
 
 // FUNCTION DEFINITIONS.
@@ -174,7 +174,7 @@ void print_leaves( node * root ) {
 		for (i = 0; i < c->num_keys; i++) {
 			if (verbose_output)
 				printf("%lx ", (unsigned long)c->pointers[i]);
-			printf("%d ", c->keys[i]);
+			printf("%zd ", c->keys[i]);
 		}
 		if (verbose_output)
 			printf("%lx ", (unsigned long)c->pointers[order - 1]);
@@ -254,7 +254,7 @@ void print_tree( node * root ) {
 		for (i = 0; i < n->num_keys; i++) {
 			if (verbose_output)
 				printf("%lx ", (unsigned long)n->pointers[i]);
-			printf("%d ", n->keys[i]);
+			printf("%zd ", n->keys[i]);
 		}
 		if (!n->is_leaf)
 			for (i = 0; i <= n->num_keys; i++)
@@ -274,12 +274,12 @@ void print_tree( node * root ) {
 /* Finds the record under a given key and prints an
  * appropriate message to stdout.
  */
-void find_and_print(node * root, int key, bool verbose) {
+void find_and_print(node * root, size_t key, bool verbose) {
 	record * r = find(root, key, -1);
 	if (r == NULL)
-		printf("Record not found under key %d.\n", key);
+		printf("Record not found under key %zd.\n", key);
 	else 
-		printf("Record at %lx -- key %d, value %s.\n",
+		printf("Record at %lx -- key %zd, value %s.\n",
 				(unsigned long)r, key, r->value);
 }
 
@@ -287,7 +287,7 @@ void find_and_print(node * root, int key, bool verbose) {
 /* Finds and prints the keys, pointers, and values within a range
  * of keys between key_start and key_end, including both bounds.
  */
-void find_and_print_range( node * root, int key_start, int key_end,
+void find_and_print_range( node * root, size_t key_start, size_t key_end,
 		bool verbose ) {
 	int i;
 	int array_size = key_end - key_start + 1;
@@ -313,7 +313,7 @@ void find_and_print_range( node * root, int key_start, int key_end,
  * returned_keys and returned_pointers, and returns the number of
  * entries found.
  */
-int find_range( node * root, int key_start, int key_end, bool verbose,
+int find_range( node * root, size_t key_start, size_t key_end, bool verbose,
 		int returned_keys[], void * returned_pointers[]) {
 	int i, num_found;
 	num_found = 0;
@@ -339,7 +339,7 @@ int find_range( node * root, int key_start, int key_end, bool verbose,
  * if the verbose flag is set.
  * Returns the leaf containing the given key.
  */
-node * find_leaf( node * root, int key, bool verbose ) {
+node * find_leaf( node * root, size_t key, bool verbose ) {
 	int i = 0;
 	node * c = root;
 	if (c == NULL) {
@@ -351,8 +351,8 @@ node * find_leaf( node * root, int key, bool verbose ) {
 		if (verbose) {
 			printf("[");
 			for (i = 0; i < c->num_keys - 1; i++)
-				printf("%d ", c->keys[i]);
-			printf("%d] ", c->keys[i]);
+				printf("%zd ", c->keys[i]);
+			printf("%zd] ", c->keys[i]);
 		}
 		i = 0;
 		while (i < c->num_keys) {
@@ -366,8 +366,8 @@ node * find_leaf( node * root, int key, bool verbose ) {
 	if (verbose) {
 		printf("Leaf [");
 		for (i = 0; i < c->num_keys - 1; i++)
-			printf("%d ", c->keys[i]);
-		printf("%d] ->\n", c->keys[i]);
+			printf("%zd ", c->keys[i]);
+		printf("%zd] ->\n", c->keys[i]);
 	}
 	return c;
 }
@@ -390,16 +390,16 @@ node * find_left_most_leaf( node * root, bool verbose ) {
 		if (verbose) {
 			printf("[");
 			for (i = 0; i < c->num_keys - 1; i++)
-				printf("%d ", c->keys[i]);
-			printf("%d] ", c->keys[i]);
+				printf("%zd ", c->keys[i]);
+			printf("%zd] ", c->keys[i]);
 		}
 		c = (node *)c->pointers[0];
 	}
 	if (verbose) {
 		printf("Leaf [");
 		for (i = 0; i < c->num_keys - 1; i++)
-			printf("%d ", c->keys[i]);
-		printf("%d] ->\n", c->keys[i]);
+			printf("%zd ", c->keys[i]);
+		printf("%zd] ->\n", c->keys[i]);
 	}
 	return c;
 }
@@ -422,16 +422,16 @@ node * find_right_most_leaf( node * root, bool verbose ) {
 		if (verbose) {
 			printf("[");
 			for (i = 0; i < c->num_keys - 1; i++)
-				printf("%d ", c->keys[i]);
-			printf("%d] ", c->keys[i]);
+				printf("%zd ", c->keys[i]);
+			printf("%zd] ", c->keys[i]);
 		}
 		c = (node *)c->pointers[c->num_keys];
 	}
 	if (verbose) {
 		printf("Leaf [");
 		for (i = 0; i < c->num_keys - 1; i++)
-			printf("%d ", c->keys[i]);
-		printf("%d] ->\n", c->keys[i]);
+			printf("%zd ", c->keys[i]);
+		printf("%zd] ->\n", c->keys[i]);
 	}
 	return c;
 }
@@ -450,7 +450,7 @@ chunk * get_left_most_leaf( node * root, bool verbose ) {
 	// Improvement can be made be allocating only num_keys elements
 	chunk *left_most_leaf=(chunk *)malloc(sizeof(chunk));
 	left_most_leaf->value=(record **)calloc(order,sizeof(record *));
-	left_most_leaf->key=(int *)malloc(order*sizeof(int));
+	left_most_leaf->key=(size_t *)malloc(order*sizeof(size_t));
 
 	int i=0;
 	for (i=0;i<c->num_keys;i++) {
@@ -475,7 +475,7 @@ chunk * get_right_most_leaf( node * root, bool verbose ) {
 	// Improvement can be made be allocating only num_keys elements
 	chunk *right_most_leaf=(chunk *)malloc(sizeof(chunk));
 	right_most_leaf->value=(record **)calloc(order,sizeof(record *));
-	right_most_leaf->key=(int *)malloc(order*sizeof(int));
+	right_most_leaf->key=(size_t *)malloc(order*sizeof(size_t));
 
 	int i=0;
 	for (i=0;i<c->num_keys;i++) {
@@ -490,7 +490,7 @@ chunk * get_right_most_leaf( node * root, bool verbose ) {
 /* Finds and returns the record to which
  * a key refers.
  */
-record * find( node * root, int key, int thread_id ) {
+record * find( node * root, size_t key, int thread_id ) {
 	// Get the current version
 	long long reader_version = global_version;
 
@@ -626,7 +626,7 @@ int get_left_index(node * parent, node * left) {
  * key into a leaf.
  * Returns the altered leaf.
  */
-node * insert_into_leaf( node * leaf, int key, record * pointer ) {
+node * insert_into_leaf( node * leaf, size_t key, record * pointer ) {
 
 	int i, insertion_point;
 
@@ -650,7 +650,7 @@ node * insert_into_leaf( node * leaf, int key, record * pointer ) {
  * the tree's order, causing the leaf to be split
  * in half.
  */
-node * insert_into_leaf_after_splitting(node * root, node * leaf, int key, record * pointer) {
+node * insert_into_leaf_after_splitting(node * root, node * leaf, size_t key, record * pointer) {
 
 	node * new_leaf;
 	int * temp_keys;
@@ -726,7 +726,7 @@ node * insert_into_leaf_after_splitting(node * root, node * leaf, int key, recor
  * without violating the B+ tree properties.
  */
 node * insert_into_node(node * root, node * n, 
-		int left_index, int key, node * right) {
+		int left_index, size_t key, node * right) {
 	int i;
 
 	for (i = n->num_keys; i > left_index; i--) {
@@ -745,7 +745,7 @@ node * insert_into_node(node * root, node * n,
  * the order, and causing the node to split into two.
  */
 node * insert_into_node_after_splitting(node * root, node * old_node, int left_index, 
-		int key, node * right) {
+		size_t key, node * right) {
 
 	int i, j, split, k_prime;
 	node * new_node, * child;
@@ -830,7 +830,7 @@ node * insert_into_node_after_splitting(node * root, node * old_node, int left_i
 /* Inserts a new node (leaf or internal node) into the B+ tree.
  * Returns the root of the tree after insertion.
  */
-node * insert_into_parent(node * root, node * left, int key, node * right) {
+node * insert_into_parent(node * root, node * left, size_t key, node * right) {
 
 	int left_index;
 	node * parent;
@@ -871,7 +871,7 @@ node * insert_into_parent(node * root, node * left, int key, node * right) {
  * and inserts the appropriate key into
  * the new root.
  */
-node * insert_into_new_root(node * left, int key, node * right) {
+node * insert_into_new_root(node * left, size_t key, node * right) {
 
 	node * root = make_node();
 	root->keys[0] = key;
@@ -889,7 +889,7 @@ node * insert_into_new_root(node * left, int key, node * right) {
 /* First insertion:
  * start a new tree.
  */
-node * start_new_tree(int key, record * pointer) {
+node * start_new_tree(size_t key, record * pointer) {
 
 	node * root = make_leaf();
 	root->keys[0] = key;
@@ -908,7 +908,7 @@ node * start_new_tree(int key, record * pointer) {
  * however necessary to maintain the B+ tree
  * properties.
  */
-node * bptinsert( node * root, int key, char *value ) {
+node * bptinsert( node * root, size_t key, char *value ) {
 
 	record * pointer;
 	node * leaf;
@@ -987,7 +987,7 @@ int get_neighbor_index( node * n ) {
 }
 
 
-node * remove_entry_from_node(node * n, int key, node * pointer) {
+node * remove_entry_from_node(node * n, size_t key, node * pointer) {
 
 	int i, num_pointers;
 
@@ -1294,7 +1294,7 @@ node * redistribute_nodes(node * root, node * n, node * neighbor, int neighbor_i
  * from the leaf, and then makes all appropriate
  * changes to preserve the B+ tree properties.
  */
-node * delete_entry( node * root, node * n, int key, void * pointer ,int worker_id) {
+node * delete_entry( node * root, node * n, size_t key, void * pointer ,int worker_id) {
 
 	int min_keys;
 	node * neighbor;
@@ -1365,7 +1365,7 @@ node * delete_entry( node * root, node * n, int key, void * pointer ,int worker_
 
 /* Master deletion function.
  */
-node * bptdelete(node * root, int key, int worker_id) {
+node * bptdelete(node * root, size_t key, int worker_id) {
 
 	node * key_leaf;
 	record * key_record;
